+++ src/app/app.module.ts
@@ -1,13 +1,16 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-
+import { Router } from '@angular/router';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
 import { CoreModule } from './core/core.module';
-import { CartModule } from './cart/cart.module';
 import { ProductsModule } from './products/products.module';
-import { AboutComponent } from './layout/components/about/about.component';
+import { LayoutModule } from './layout/layout.module';
+import { AboutComponent } from './layout';
+import { ProductsRoutingModule } from './products/products-routing.module';
+import { OrderModule } from './order/order.module';
+import { AdminModule } from './admin/admin.module';
 
 
 @NgModule({
@@ -20,11 +23,21 @@ import { AboutComponent } from './layout/components/about/about.component';
     BrowserModule,
     SharedModule,
     CoreModule,
-    CartModule,
     ProductsModule,
-    AppRoutingModule // <-- старайтесь, чтобы этот модуль был последним
+    LayoutModule,
+    OrderModule,
+    ProductsRoutingModule,
+    AppRoutingModule,
+    AdminModule
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule {}
+export class AppModule {
+  constructor(router: Router) {
+    const replacer = (key: string, value: any): string =>
+      typeof value === 'function' ? value.name : value;
+
+    console.log('Routes: ', JSON.stringify(router.config, replacer, 2));
+  }
+}