+++ src/app/cart/cart.module.ts
@@ -2,19 +2,25 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SharedModule } from '../shared/shared.module';
 import { CartComponent } from './components/cart/cart.component';
-import { CartItemComponent } from './components/cart-item/cart-item.component';
+import { CartRoutingModule } from './cart-routing.module';
+import { RouterModule } from '@angular/router';
+import { CartContainerComponent } from './cart-container.component';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [
-    CartComponent,
-    CartItemComponent
+    CartRoutingModule.components
   ],
   imports: [
     CommonModule,
-    SharedModule
+    SharedModule,
+    FormsModule,
+    RouterModule,
+    CartRoutingModule
   ],
   exports: [
-    CartComponent
+    CartComponent,
+    CartContainerComponent
   ]
 })
 export class CartModule { }