+++ src/app/layout/components/login/login.component.ts
@@ -1,0 +1,59 @@
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Router } from '@angular/router';
+
+// rxjs
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { AuthService } from 'src/app/core/services/auth.service';
+
+@Component({
+  selector: 'app-login',
+  templateUrl: './login.component.html',
+  styleUrls: ['./login.component.css']
+})
+export class LoginComponent implements OnInit, OnDestroy {
+  message: string;
+  private unsubscribe: Subject<void> = new Subject();
+
+  constructor(
+    public authService: AuthService,
+    private router: Router
+  ) { }
+
+  ngOnInit() {
+    this.setMessage();
+  }
+
+  ngOnDestroy() {
+    console.log('[takeUntil ngOnDestroy]');
+    this.unsubscribe.next();
+    this.unsubscribe.complete();
+  }
+  onLogin() {
+    this.message = 'Trying to log in ...';
+    this.authService
+      .login()
+      .pipe(takeUntil(this.unsubscribe))
+      .subscribe(
+       () => {
+         this.setMessage();
+         if (this.authService.isLoggedIn) {
+          const redirect = this.authService.redirectUrl ? this.authService.redirectUrl : 'admin';
+          this.router.navigate([redirect]);
+      }
+    },
+        err => console.log(err),
+        () => console.log('[takeUntil] complete')
+    );
+  }
+
+onLogout() {
+    this.authService.logout();
+    this.setMessage();
+}
+
+private setMessage() {
+    this.message = 'Logged ' + (this.authService.isLoggedIn ? 'in' : 'out');
+}
+
+}