+++ src/app/layout/layout.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { AboutComponent, LoginComponent, PathNotFoundComponent } from './components';
+import { BrowserModule } from '@angular/platform-browser';
+
+@NgModule({
+  declarations: [
+    PathNotFoundComponent,
+    LoginComponent
+  ],
+  imports: [
+    BrowserModule
+  ],
+  exports: [
+    PathNotFoundComponent
+  ]
+
+})
+export class LayoutModule { }